refactor: drop _pyg suffix from GRACE modules and fix mypy + parity CI
- Rename:
    src/matgl/layers/_grace_pyg.py    -> src/matgl/layers/_grace.py
    src/matgl/models/_grace_pyg.py    -> src/matgl/models/_grace.py
    tests/layers/test_grace_pyg.py    -> tests/layers/test_grace.py

  Updates all internal imports and the per-file ruff E741 ignore list to
  the new module names. The parity test (tests/models/test_grace_parity_tp.py)
  is unchanged in name (not GRACE-implementation, kept distinct) but its
  import path is updated.

- Fix mypy error in LinearRadialFunction.forward:
    src/matgl/layers/_grace_pyg.py:126: error: Argument "index" to
    "index_select" of "TensorBase" has incompatible type "Tensor | Module";
    expected "Tensor"  [arg-type]

  Buffers registered via nn.Module.register_buffer have static type
  Tensor | Module, so .index_select(index=self._l_idx) trips mypy. We
  narrow the access with torch.as_tensor(self._l_idx), matching the
  existing pattern in matgl/layers/_so3.py SO3Convolution.

- Fix Linux test_grace_parity CI job:
    On Linux, tensorpotential 0.5.9 pins tensorflow[and-cuda]<2.20, whose
    nvidia-nccl-cu12 extra clashes with torch's bundled NCCL ABI:
        ImportError: libtorch_cuda.so: undefined symbol: ncclCommWindowDeregister
    Add a uv `--override` constraint to install plain tensorflow<2.20
    (CPU-only). The parity tests are deterministic CPU math and don't
    need the GPU stack.
