+++ src/matgl/models/__init__.py
@@ -14,7 +14,7 @@ if BACKEND == "DGL":
     from ._so3net import SO3Net
     from ._tensornet_dgl import TensorNet
 else:
-    from ._grace_pyg import GRACE
+    from ._grace import GRACE
     from ._m3gnet_pyg import M3GNet  # type: ignore[assignment]
     from ._megnet_pyg import MEGNet  # type: ignore[assignment]
     from ._qet_pyg import QET  # type: ignore[assignment]
